Adding __invoke to support closures
In Traq\Installer\Controllers\AppController you set installStep as a closure.  Then in license_agreement.phtml you have the form action as $installStep("database_info") which fails with a fatal error stating that "Function name must be a string".  A var_dump shows that the $installStep var is actually the closure wrapped in a SafeObject but that it doesn't route correctly to the enclosed closure.  By adding the __invoke function we can allow a SafeObject to be used as a closure and function as expected.